% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binomial_LocalisedINLA.R
\name{binomial_localisedINLA}
\alias{binomial_localisedINLA}
\title{Run a Binomial Localised spatial autoregressive model using INLA}
\usage{
binomial_localisedINLA(
  formula,
  W,
  Ntrials,
  fix.rho = FALSE,
  rho = NULL,
  prior.beta.mean = 0,
  prior.beta.precision = 0.001,
  prior.precision.shape = 0.001,
  prior.precision.scale = 0.001
)
}
\description{
Run a Binomial Localised spatial autoregressive model using INLA
}
