% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frontier_detect.R
\name{frontier_detect}
\alias{frontier_detect}
\title{This is a wrapper for our frontier detection functions
goal: standardise input and outputs regardless of technique?
TODO: Check out how packages like matchit and synthpop deal with calling
methods from other packages
 Note: orginally called inla_frontier in the project
 Returns list object of class: frontier_model}
\usage{
frontier_detect(y, data, n.trials, W.nb = NULL, ...)
}
\description{
This is a wrapper for our frontier detection functions
goal: standardise input and outputs regardless of technique?
TODO: Check out how packages like matchit and synthpop deal with calling
methods from other packages
 Note: orginally called inla_frontier in the project
 Returns list object of class: frontier_model
}
