% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/london.R
\docType{data}
\name{london}
\alias{london}
\title{London data [Placeholder]}
\format{
LSOA:
Areal unit code

LSOAname Areal unit name. The first part denotes the local authority (e.g. Barnet).

UK Total count of UK born

nonUK Total count of non-UK born

totalPop Total resident population in LSOA

econActive Total resident economically active (not necessarily employed)

unemployed Total residents unemployed

totalcrimes Total crime count from police.co.uk data
}
\usage{
london
}
\description{
This is the data from London as an sf object with spatial polygons. Each record
denotes a lower super output area (LSOA) in London.
}
\details{
Originally sourced and formatted by Dan Olner.
}
\keyword{datasets}
