% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_summary_frontier_model.R
\name{print.summary.frontier_model}
\alias{print.summary.frontier_model}
\title{Print method for summary.frontier_model}
\usage{
\method{print}{summary.frontier_model}(object, ...)
}
\description{
See summary.frontier_model
}
