% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startMyProject.R
\name{startMyProject}
\alias{startMyProject}
\title{The very first function in the entire project that loads in where the
 data is and packages that we need}
\usage{
startMyProject()
}
\description{
The very first function in the entire project that loads in where the
 data is and packages that we need
}
