% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_frontier_model.R
\name{summary.frontier_model}
\alias{summary.frontier_model}
\title{Common inspections of the object include the number of frontiers vs non-frontiers
this uses a class method for summary -- summary.<class> happens if we do 
summary(<this_class>)}
\usage{
\method{summary}{frontier_model}(object, ...)
}
\description{
Common inspections of the object include the number of frontiers vs non-frontiers
this uses a class method for summary -- summary.<class> happens if we do 
summary(<this_class>)
}
